feat(company subscriptions): fixed review comments
