+++ src/assets/locales/de/main.json
@@ -65,7 +65,7 @@
     "admin-credential": "Certification Mgmt",
     "companyCertificate": "Company Certificates",
     "companyWallet": "Company Wallet",
-    "company-subscriptions": "vFirmenabonnements",
+    "company-subscriptions": "Firmenabonnements",
     "mycompany": "My Company",
     "mynotifications": "My Notifications"
   },
