+++ src/components/pages/CompanySubscriptions/CompanySubscriptionsTableColumns.tsx
@@ -1,0 +1,172 @@
+/********************************************************************************
+ * Copyright (c) 2021, 2024 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ ********************************************************************************/
+
+import {
+  LogoGrayData,
+  Image,
+  IconButton,
+  Button,
+  Typography,
+} from '@catena-x/portal-shared-components'
+import type { GridColDef } from '@mui/x-data-grid'
+import {
+  type SubscribedActiveApps,
+  SubscriptionStatus,
+} from 'features/apps/types'
+import { getApiBase } from 'services/EnvironmentService'
+import { fetchImageWithToken } from 'services/ImageService'
+import { PAGES } from 'types/Constants'
+import { useNavigate } from 'react-router'
+import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline'
+import UnpublishedIcon from '@mui/icons-material/Unpublished'
+import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
+import type i18next from 'i18next'
+import { Box } from '@mui/material'
+
+export const CompanySubscriptionsTableColumns = (
+  t: typeof i18next.t,
+  handleOverlay?: (row: SubscribedActiveApps, enable: boolean) => void
+): Array<GridColDef> => {
+  const navigate = useNavigate()
+
+  const renderStatus = (status: string) => {
+    if (status === SubscriptionStatus.ACTIVE)
+      return (
+        <Button
+          variant="text"
+          startIcon={<CheckCircleOutlineIcon />}
+          sx={{ color: '#B3CB2D', pointerEvents: 'none' }}
+          size="small"
+        >
+          {t('content.companySubscriptions.subscribed')}
+        </Button>
+      )
+    else if (status === SubscriptionStatus.PENDING)
+      return (
+        <Button
+          variant="outlined"
+          sx={{
+            color: '#FFA600',
+            borderColor: '#FFA600',
+            pointerEvents: 'none',
+            ml: 1,
+          }}
+          size="small"
+        >
+          {t('content.companySubscriptions.requested')}
+        </Button>
+      )
+    else
+      return (
+        <Button
+          variant="text"
+          startIcon={<UnpublishedIcon />}
+          sx={{ color: '#D91E18', pointerEvents: 'none' }}
+          size="small"
+        >
+          {t('content.companySubscriptions.declined')}
+        </Button>
+      )
+  }
+
+  return [
+    {
+      field: 'image',
+      headerName: t('content.companySubscriptions.table.appIcon'),
+      flex: 2,
+      renderCell: ({ row }: { row: SubscribedActiveApps }) => (
+        <Image
+          src={
+            row.image
+              ? `${getApiBase()}/api/apps/${row.offerId}/appDocuments/${
+                  row.image
+                }`
+              : LogoGrayData
+          }
+          loader={fetchImageWithToken}
+          style={{
+            objectFit: 'cover',
+            width: 30,
+            height: 30,
+            borderRadius: '50%',
+            marginRight: '5px',
+          }}
+        />
+      ),
+    },
+    {
+      field: 'name',
+      headerName: t('content.companySubscriptions.table.appTitle'),
+      flex: 3,
+      renderCell: ({ row }: { row: SubscribedActiveApps }) => (
+        <Box sx={{ display: 'grid' }}>
+          <Typography variant="body3">{row.name}</Typography>
+          <Typography variant="body3">{row.provider}</Typography>
+        </Box>
+      ),
+    },
+    {
+      field: 'status',
+      headerName: t('content.companySubscriptions.table.status'),
+      flex: 2,
+      renderCell: ({ row }: { row: SubscribedActiveApps }) => {
+        return renderStatus(row.status)
+      },
+    },
+    {
+      field: 'details',
+      headerName: t('content.companySubscriptions.table.details'),
+      flex: 2,
+      renderCell: ({ row }: { row: SubscribedActiveApps }) => {
+        return (
+          <IconButton
+            color="secondary"
+            onClick={() => {
+              navigate(
+                `/${PAGES.COMPANY_SUBSCRIPTIONS_DETAIL}/${row.offerId}`,
+                { state: row }
+              )
+            }}
+          >
+            <ArrowForwardIcon />
+          </IconButton>
+        )
+      },
+    },
+    {
+      field: 'action',
+      headerName: t('content.companySubscriptions.table.action'),
+      flex: 2,
+      renderCell: ({ row }: { row: SubscribedActiveApps }) =>
+        row.status === SubscriptionStatus.ACTIVE && (
+          <Button
+            variant="contained"
+            size="small"
+            onClick={(e) => {
+              handleOverlay?.(row, true)
+              console.log('row.offerId', row.offerId)
+              e.stopPropagation()
+            }}
+          >
+            {t('content.companySubscriptions.unsubscribe')}
+          </Button>
+        ),
+    },
+  ]
+}
